pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "mmpp-soccer"
version = "0.1.0"
description = "MMPP Soccer Live Trading System"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26,<2.0",
    "numba>=0.59,<1.0",
    "scipy>=1.12,<2.0",
    "xgboost>=2.0,<3.0",
    "torch>=2.2,<3.0",
    "httpx>=0.27,<1.0",
    "websockets>=12.0,<14.0",
    "asyncpg>=0.29,<1.0",
    "redis>=5.0,<6.0",
    "structlog>=24.1,<25.0",
    "prometheus-client>=0.20,<1.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "hypothesis>=6.98,<7.0",
    "mypy>=1.8,<2.0",
    "ruff>=0.3,<1.0",
    "types-PyYAML>=6.0,<7.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
ignore = ["E501", "N815", "TC003"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "numba.*",
    "xgboost.*",
    "hypothesis.*",
    "asyncpg.*",
    "redis.*",
    "prometheus_client.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: integration tests (mock APIs)",
    "property: hypothesis property-based tests",
]
